requirements.txt
tensorflow-lite==2.5.0
opencv-python==4.5.3.56
numpy==1.19.5
pillow==8.3.1
scikit-image==0.18.2
flask==2.0.1